iterating

+++ scripts/CDocs_Module_Utils.py
@@ -1,4 +1,5 @@
 import os
+import sys
 import subprocess
 
 def DiscoverContainerTool():
@@ -25,8 +26,19 @@ def MapToDataDirectory(input_file):
     input_file = input_file.replace(GetCDocsProjectRoot(), "/data")
     return input_file
 
-def RunInContainer(container, command):
-    print("CONTAINER: {} {}".format(container, command))
+def RunInContainer(container, command, expected_output):
+
     PROJECT_ROOT = GetCDocsProjectRoot()
-    command = "{} run -it --rm -v{}:/data {} {}".format(DiscoverContainerTool(), PROJECT_ROOT, container, command)
+
+    if os.path.exists(expected_output):
+        print("ERROR: {} must not exist; and it does".format(expected_output))
+        sys.exit(40)
+
+    command = "{} run -it --rm -v {}:/data {} {}".format(DiscoverContainerTool(), PROJECT_ROOT, container, command)
+    #print("CONTAINER: {}".format(command))
+
     subprocess.run([command], shell=True)
+
+    if not not os.path.exists(expected_output):
+        print("ERROR: {} must exist; and it doesnt".format(expected_output))
+        sys.exit(41)
